Working on OAuth authentication...
I started making requests to collections and got some authentication
errors. So, after working for an our on the depreciated authentication
protocol, I started in on OAuth. Which is super complicated. Long story
short this commit is very unstable and definitely broken, I wouldn't
pull this one unless you're developing.